separate ledger accounts from hardware management
