+++ src/contexts/Hardware/Ledger/static/ledger.ts
