+++ src/contexts/Hardware/Ledger/types.ts
@@ -1,7 +1,7 @@
 // Copyright 2023 @paritytech/polkadot-staking-dashboard authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-import type { LedgerAccount, VaultAccount } from '@polkadot-cloud/react/types';
+import type { LedgerAccount } from '@polkadot-cloud/react/types';
 import type { FunctionComponent, SVGProps } from 'react';
 import type { AnyJson, MaybeString, NetworkName } from 'types';
 
@@ -15,12 +15,6 @@ export type LedgerHardwareContextInterface = {
   resetStatusCodes: () => void;
   getIsExecuting: () => boolean;
   getStatusCodes: () => LedgerResponse[];
-  ledgerAccountExists: (a: string) => boolean;
-  addLedgerAccount: (a: string, i: number) => LedgerAccount | null;
-  removeLedgerAccount: (a: string) => void;
-  renameLedgerAccount: (a: string, name: string) => void;
-  getLedgerAccount: (a: string) => LedgerAccount | null;
-  ledgerAccounts: LedgerAccount[];
   getFeedback: () => FeedbackMessage;
   setFeedback: (s: MaybeString, helpKey?: MaybeString) => void;
   resetFeedback: () => void;
@@ -35,6 +29,15 @@ export type LedgerHardwareContextInterface = {
     payload: AnyJson
   ) => Promise<void>;
   handleResetLedgerTx: () => void;
+};
+
+export type LedgerAccountsContextInterface = {
+  ledgerAccountExists: (a: string) => boolean;
+  addLedgerAccount: (a: string, i: number) => LedgerAccount | null;
+  removeLedgerAccount: (a: string) => void;
+  renameLedgerAccount: (a: string, name: string) => void;
+  getLedgerAccount: (a: string) => LedgerAccount | null;
+  ledgerAccounts: LedgerAccount[];
 };
 
 export interface FeedbackMessage {
@@ -80,13 +83,4 @@ export type LedgerApp = {
   network: NetworkName;
   appName: string;
   Icon: FunctionComponent<SVGProps<SVGSVGElement>>;
-};
-
-export type VaultHardwareContextInterface = {
-  vaultAccountExists: (a: string) => boolean;
-  addVaultAccount: (a: string, i: number) => LedgerAccount | null;
-  removeVaultAccount: (a: string) => void;
-  renameVaultAccount: (a: string, name: string) => void;
-  getVaultAccount: (a: string) => LedgerAccount | null;
-  vaultAccounts: VaultAccount[];
 };
