+++ src/library/SubmitTx/ManualSign/Ledger/index.tsx
@@ -5,8 +5,8 @@ import { ButtonHelp, useEffectIgnoreInitial } from '@polkadot-cloud/react';
 import type { ReactNode } from 'react';
 import { useEffect, useRef } from 'react';
 import { useTranslation } from 'react-i18next';
-import { useLedgerHardware } from 'contexts/Hardware/Ledger';
-import type { LedgerResponse } from 'contexts/Hardware/types';
+import { useLedgerHardware } from 'contexts/Hardware/Ledger/LedgerHardware';
+import type { LedgerResponse } from 'contexts/Hardware/Ledger/types';
 import { useHelp } from 'contexts/Help';
 import { useTxMeta } from 'contexts/TxMeta';
 import { EstimatedTxFee } from 'library/EstimatedTxFee';
