+++ src/modals/ImportLedger/Reset.tsx
@@ -3,22 +3,22 @@
 
 import { ButtonMono, ButtonMonoInvert } from '@polkadot-cloud/react';
 import { useTranslation } from 'react-i18next';
-import { useLedgerHardware } from 'contexts/Hardware/Ledger';
 import { getLocalLedgerAddresses } from 'contexts/Hardware/Utils';
-import type { LedgerAddress } from 'contexts/Hardware/types';
+import type { LedgerAddress } from 'contexts/Hardware/Ledger/types';
 import { usePrompt } from 'contexts/Prompt';
 import { ConfirmWrapper } from 'library/Import/Wrappers';
 import type { AnyJson } from 'types';
 import { useOverlay } from '@polkadot-cloud/react/hooks';
 import { useOtherAccounts } from 'contexts/Connect/OtherAccounts';
 import type { LedgerAccount } from '@polkadot-cloud/react/types';
+import { useLedgerAccounts } from 'contexts/Hardware/Ledger/LedgerAccounts';
 
 export const Reset = ({ removeLedgerAddress }: AnyJson) => {
   const { t } = useTranslation('modals');
   const { setStatus } = usePrompt();
   const { replaceModal } = useOverlay().modal;
   const { forgetOtherAccounts } = useOtherAccounts();
-  const { ledgerAccounts, removeLedgerAccount } = useLedgerHardware();
+  const { removeLedgerAccount, ledgerAccounts } = useLedgerAccounts();
 
   const removeAccounts = () => {
     // Remove imported Ledger accounts.
